Pused updated kernel, EQ not working as of now

+++ Sources/audioKernel_0.3/spimctrl.h
@@ -9,5 +9,5 @@ void spiFlashWrite(uint8_t addressArray[3], uint8_t* buf, int bytes);
 void spiFlashSectorErase(uint8_t addressArray[3]);
 void spiFlashRead(uint8_t addressArray[3],uint8_t* buf,int bytes);
 
-void setPreset(uint8_t id, uint8_t* buf);
-void getPreset(uint8_t id,uint8_t* buf);
+void setPreset(uint8_t id, int32_t* buf);
+void getPreset(uint8_t id,int32_t* buf);
